+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 0.0.20
   - Handle for non content type responses
-  - Bug fix for GetMyPriceForASIN and similer other products API
+  - Bug fix for GetMyPriceForASIN and similar other products API
   - Add example, test case for GetMyPriceForASIN
   - Add test case for failure/error responses