+++ CHANGELOG.md
@@ -3,6 +3,9 @@
   - Bug fix for GetMyPriceForASIN and similar other products API
   - Add example, test case for GetMyPriceForASIN
   - Add test case for failure/error responses
+  - Add support for RequestReport API
+  - Add example for GetMyFeesEstimate in Products
+  - Add force check for the XML string
   
 0.0.19
   - Correct lint and increase test case timeout